sshca: support method token and handle multi line (#109)
We didn't account that the SSH_AUTH_INFO_0 variable format has a method
token and that it can have multiple lines - even that it's unlikely
to happen with oslogin use cases it's healthier to account for that.

The tests were changed to reflect the actual implementation change.

Additionally this patch also changes the internal function's signature
to be prefixed with _.