byoid: add a pam module to check for byoid cert fingerprint
BYOID authentication is a first in stack auth module set as sufficient,
meaning that it will be ignored if failed and no other auth module will
be checked if succeeded. The implementation accounts for 2fa module and
certificate based authentication.

if it's a cert based auth and it contains a BYOID fingerprint extension
then we return success (ignoring other auth modules), otherwise it will
return failure if 2fa is enabled or success if 2fa is disabled.